migrate change to default menu
