+++ src/components/AccountDrawer/DefaultMenu.tsx
@@ -1,6 +1,7 @@
 import { useWeb3React } from '@web3-react/core'
 import Column from 'components/Column'
 import WalletModal from 'components/WalletModal'
+import { useCallback, useEffect, useState } from 'react'
 import styled from 'styled-components/macro'
 
 import AuthenticatedHeader from './AuthenticatedHeader'
@@ -16,17 +17,22 @@ enum MenuState {
   SETTINGS,
 }
 
-function DefaultMenu({
-  menu,
-  openSettings,
-  closeSettings,
-}: {
-  menu: MenuState
-  openSettings: () => void
-  closeSettings: () => void
-}) {
+function DefaultMenu({ drawerOpen }: { drawerOpen: boolean }) {
   const { account } = useWeb3React()
   const isAuthenticated = !!account
+
+  const [menu, setMenu] = useState<MenuState>(MenuState.DEFAULT)
+  const openSettings = useCallback(() => setMenu(MenuState.SETTINGS), [])
+  const closeSettings = useCallback(() => setMenu(MenuState.DEFAULT), [])
+
+  useEffect(() => {
+    if (!drawerOpen && menu === MenuState.SETTINGS) {
+      // wait for the drawer to close before resetting the menu
+      setTimeout(() => {
+        closeSettings()
+      }, 250)
+    }
+  }, [drawerOpen, menu, closeSettings])
 
   return (
     <DefaultMenuWrap>
